Create usePromptWithAnswer hook
