+++ src/types/behavioural-interviews.ts
@@ -2,12 +2,12 @@ export type Topic = {
   id: number;
   title: string;
   order: number;
-}
+};
 
 export enum PromptType {
   Freeform = "freeform",
   Star = "star",
-}
+};
 
 export type Prompt = {
   id: number;
@@ -15,30 +15,25 @@ export type Prompt = {
   type: PromptType;
   topicId: number; // references Topic
   order: number; // order within a topic
-}
+};
 
 export type TopicWithPrompts = Topic & {
   prompts: Prompt[];
-}
+};
 
-type BasePromptAnswer = {
+export type FreeformAnswer = {
+  text: string;
+};
+
+export type StarAnswer = {
+  situation: string;
+  task: string;
+  action: string;
+  result: string;
+};
+
+export type PromptAnswer = {
   id: number;
   promptId: number; // references Prompt
-}
-
-export type FreeformPromptAnswer = BasePromptAnswer & {
-  answer: {
-    text: string;
-  }
-}
-
-export type StarPromptAnswer = BasePromptAnswer & {
-  answer: {
-    situation: string;
-    task: string;
-    action: string;
-    result: string;
-  }
-}
-
-export type PromptAnswer = FreeformPromptAnswer | StarPromptAnswer;
+  answer: FreeformAnswer | StarAnswer;
+};
